# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quran Memorization Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rewards/prizes:
    get:
      operationId: listRewardPrizes
      tags: [rewards]
      summary: List the user's prizes
      responses:
        "200":
          description: Prize list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RewardPrize"
    post:
      operationId: createRewardPrize
      tags: [rewards]
      summary: Create a prize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRewardPrizeBody"
      responses:
        "201":
          description: Created prize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardPrize"
  /rewards/prizes/{id}:
    patch:
      operationId: updateRewardPrize
      tags: [rewards]
      summary: Update a prize
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRewardPrizeBody"
      responses:
        "200":
          description: Updated prize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardPrize"
        "404":
          description: Prize not found
    delete:
      operationId: deleteRewardPrize
      tags: [rewards]
      summary: Delete a prize
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Prize not found
  /rewards/prizes/{id}/redeem:
    post:
      operationId: redeemRewardPrize
      tags: [rewards]
      summary: Redeem (collect) a prize using earned points
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Redemption recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemRewardPrizeResult"
        "400":
          description: Insufficient points
        "404":
          description: Prize not found
...
components:
  schemas:
...
    RewardPrize:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cost:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, cost, createdAt]
    CreateRewardPrizeBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        cost:
          type: integer
          minimum: 1
          maximum: 100000
      required: [name, cost]
    UpdateRewardPrizeBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        cost:
          type: integer
          minimum: 1
          maximum: 100000
    RewardRedemption:
      type: object
      properties:
        id:
          type: integer
        prizeId:
          type: integer
          nullable: true
        prizeName:
          type: string
        cost:
          type: integer
        redeemedAt:
          type: string
          format: date-time
      required: [id, prizeName, cost, redeemedAt]
    RedeemRewardPrizeResult:
      type: object
      properties:
        redemption:
          $ref: "#/components/schemas/RewardRedemption"
        balance:
          type: integer
          description: Remaining balance after the redemption.
      required: [redemption, balance]
...