# lib/api-zod/src/generated/types/createRewardPrizeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Memorization Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRewardPrizeBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minimum 1
   * @maximum 100000
   */
  cost: number;
}