# lib/api-zod/src/generated/types/redeemRewardPrizeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Memorization Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RewardRedemption } from "./rewardRedemption";

export interface RedeemRewardPrizeResult {
  redemption: RewardRedemption;
  /** Remaining balance after the redemption. */
  balance: number;
}