# lib/api-zod/src/generated/types/rewardPrize.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Memorization Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface RewardPrize {
  id: number;
  name: string;
  cost: number;
  createdAt: Date;
}