# lib/api-zod/src/generated/types/rewardRedemption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Memorization Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface RewardRedemption {
  id: number;
  prizeId?: number | null;
  prizeName: string;
  cost: number;
  redeemedAt: Date;
}